Added new package build script, rebuild all packages
This improves the package build process a bit from bespoke scripts per
package. We now have a standard package build script format, and a script
that will build a package using that. It supports source downloading and
creates the package in the right location in the repo. This takes some of
the work out of rebuilding packages.

Also, rebuild all packages now that we've enabled library versioning on
system libraries.